Added all prototypes & most of Part 1

+++ ft_strlen.c
@@ -6,13 +6,15 @@
 /*   By: cfeijoo <cfeijoo@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2013/11/22 01:34:43 by cfeijoo           #+#    #+#             */
-/*   Updated: 2013/11/22 01:37:44 by cfeijoo          ###   ########.fr       */
+/*   Updated: 2013/11/24 02:09:36 by cfeijoo          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-int		ft_strlen (char *str1)
+#include <libft.h>
+
+size_t	ft_strlen (char *str1)
 {
-	int	length;
+	size_t	length;
 
 	length = 0;
 	while (str1[length])
